# docs/swagger.yaml
basePath: /
definitions:
  models.Money:
    properties:
      amount:
        type: integer
      date:
        type: string
      fromuserid:
        type: integer
      touserid:
        type: integer
    type: object
  models.Replenishment:
    properties:
      amount:
        type: integer
      date:
        type: string
      userid:
        type: integer
    type: object
  models.Response:
    properties:
      message:
        type: string
    type: object
  models.Transaction:
    properties:
      amount:
        type: integer
      date:
        type: string
      orderid:
        type: integer
      serviceid:
        type: integer
      userid:
        type: integer
    type: object
  models.User:
    properties:
      balance:
        type: integer
      userid:
        type: integer
    type: object
host: localhost:8081
info:
  contact: {}
  description: Microservice for working with user balance
  title: crtexBalance API
  version: "1.0"
paths:
  /:
    post:
      consumes:
      - application/json
      description: getting the user's balance
      operationId: get-balance
      parameters:
      - description: user id
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get Balance
      tags:
      - balance
  /topup:
    post:
      consumes:
      - application/json
      description: replenishment of the user's balance
      operationId: replenishment-balance
      parameters:
      - description: replenishment information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Replenishment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Replenishment Balance
      tags:
      - balance
  /transfer:
    post:
      consumes:
      - application/json
      description: money transfer between users
      operationId: transfer
      parameters:
      - description: transfer information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Money'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Money transfer
      tags:
      - balance
swagger: "2.0"